refactored SignedTransaction, SignedDelegateAction
